+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 - Added new `Rosenbrock` integrator.
 - Clarified naming conventions for macros.
 - Added `dt_fixed`, `rel_tol`, and `abs_tol` options to phasor dynamics solver JSON files and renamed the `dt` option to `dt_monitor`.
+- Added `closed` parameter to `Branch` for declaring out-of-service lines in case files.
 
 ## v0.1